Address PR review feedback for wildcard allowlist matching.
Register allowed-issuers and audiences in ServerProperties.Property, make
WildcardAllowlist exact entries immutable, improve audience debug logging,
document DNS-label wildcard semantics in server.properties, and consolidate
AuthService allowlist integration tests into a single nested test class.
