+++ etc/conf/server.properties
@@ -16,13 +16,15 @@ server.client-secret=
 server.redirect-port=
 
 # Required when authorization is enabled: Comma-separated list of allowed token issuers
-# (exact match or wildcard with *). Tokens from issuers not in this list will be rejected
+# (exact match or wildcard with *; each * matches a single DNS label). Tokens from issuers
+# not in this list will be rejected
 # Example for Google: https://accounts.google.com
 # Example for Microsoft Entra ID: https://login.microsoftonline.com/{tenant-id}/v2.0
 server.allowed-issuers=
 
 # Required when authorization is enabled: Comma-separated list of expected JWT audiences
-# (exact match or wildcard with *). Use "*" alone to disable audience validation.
+# (exact match or wildcard with *; each * matches a single DNS label). Use "*" alone to
+# disable audience validation.
 # Tokens not intended for these audiences will be rejected
 # Typically your application's client ID or API identifier
 server.audiences=
