--- server/src/test/java/io/unitycatalog/server/service/AuthServiceWildcardAudienceTest.java
@@ -1,110 +1,0 @@
-package io.unitycatalog.server.service;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-import com.auth0.jwt.JWT;
-import com.auth0.jwt.algorithms.Algorithm;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.linecorp.armeria.client.WebClient;
-import com.linecorp.armeria.common.AggregatedHttpResponse;
-import com.linecorp.armeria.common.HttpData;
-import com.linecorp.armeria.common.HttpMethod;
-import com.linecorp.armeria.common.HttpStatus;
-import com.linecorp.armeria.common.MediaType;
-import com.linecorp.armeria.common.RequestHeaders;
-import io.unitycatalog.server.base.auth.BaseAuthCRUDTest;
-import java.io.IOException;
-import java.util.Date;
-import java.util.UUID;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-public class AuthServiceWildcardAudienceTest extends BaseAuthCRUDTest {
-
-  private static final String TOKEN_ENDPOINT = "/api/1.0/unity-control/auth/tokens";
-  private static final ObjectMapper MAPPER = new ObjectMapper();
-
-  private WebClient client;
-
-  @Override
-  protected void setUpProperties() {
-    super.setUpProperties();
-    serverProperties.setProperty("server.audiences", "*");
-  }
-
-  @BeforeEach
-  @Override
-  public void setUp() {
-    super.setUp();
-    client = WebClient.builder(serverConfig.getServerUrl()).build();
-  }
-
-  @Test
-  public void testTokenExchangeWithWildcardAudiencesAcceptsWrongAudience() throws IOException {
-    String token =
-        createIdentityToken(
-            testIssuer, "dynamic-oauth-client-uuid", testIssuerAlgorithm, testIssuerKeyId);
-
-    AggregatedHttpResponse response = exchangeToken(token);
-
-    assertThat(response.status()).isEqualTo(HttpStatus.OK);
-    JsonNode body = MAPPER.readTree(response.contentUtf8());
-    assertThat(body.has("access_token")).isTrue();
-    assertThat(body.get("access_token").asText()).isNotEmpty();
-  }
-
-  @Test
-  public void testTokenExchangeWithWildcardAudiencesRejectsUnknownUser() {
-    String token =
-        createIdentityTokenForSubject(
-            testIssuer,
-            "any-audience",
-            "unknown-user@example.com",
-            testIssuerAlgorithm,
-            testIssuerKeyId);
-
-    AggregatedHttpResponse response = exchangeToken(token);
-
-    assertThat(response.status()).isEqualTo(HttpStatus.BAD_REQUEST);
-    assertThat(response.contentUtf8()).contains("User not allowed");
-  }
-
-  private String createIdentityToken(
-      String issuer, String audience, Algorithm algorithm, String keyId) {
-    return createIdentityTokenForSubject(issuer, audience, "admin", algorithm, keyId);
-  }
-
-  private String createIdentityTokenForSubject(
-      String issuer, String audience, String subject, Algorithm algorithm, String keyId) {
-    var builder =
-        JWT.create()
-            .withSubject(subject)
-            .withIssuer(issuer)
-            .withIssuedAt(new Date())
-            .withKeyId(keyId)
-            .withJWTId(UUID.randomUUID().toString());
-    if (audience != null) {
-      builder.withAudience(audience);
-    }
-    return builder.sign(algorithm);
-  }
-
-  private AggregatedHttpResponse exchangeToken(String identityToken) {
-    String formBody =
-        "grant_type=urn:ietf:params:oauth:grant-type:token-exchange"
-            + "&requested_token_type=urn:ietf:params:oauth:token-type:access_token"
-            + "&subject_token_type=urn:ietf:params:oauth:token-type:id_token"
-            + "&subject_token="
-            + identityToken;
-
-    RequestHeaders headers =
-        RequestHeaders.builder()
-            .method(HttpMethod.POST)
-            .path(TOKEN_ENDPOINT)
-            .contentType(MediaType.FORM_DATA)
-            .build();
-
-    return client.execute(headers, HttpData.ofUtf8(formBody)).aggregate().join();
-  }
-}
